+++ client/css/style.css
@@ -298,6 +298,7 @@ p {
 #connect .extra-help,
 #settings .extra-help,
 #settings #play::before,
+#settings .tabs .icon::before,
 #form #upload::before,
 #form #submit::before,
 #chat .msg[data-type="away"] .from::before,
@@ -397,7 +398,9 @@ p {
 }
 
 #footer .connect::before { content: "\f067"; /* http://fontawesome.io/icon/plus/ */ }
+
 #footer .settings::before { content: "\f013"; /* http://fontawesome.io/icon/cog/ */ }
+
 #footer .help::before { content: "\f059"; /* http://fontawesome.io/icon/question/ */ }
 
 #form #upload::before { content: "\f0c6"; /* https://fontawesome.com/icons/paperclip?style=solid */ }
@@ -722,7 +725,7 @@ background on hover (unless active) */
 	background-color: rgb(48 62 74 / 50%); /* #303e4a x 50% alpha */
 }
 
-/* Darker background and defualt cursor for active channels */
+/* Darker background and default cursor for active channels */
 #footer button.active,
 .channel-list-item.active {
 	background-color: #303e4a;
